API reference

Entry points

ColPack.ColPackColoring — Type

ColPackColoring

Struct holding the parameters of a coloring as well as its results (which can be queried with get_colors).

Fields

The fields of this struct are not part of the public API, they are only useful to interface with the C++ library ColPack.

Constructors

ColPackColoring(
    filename::AbstractString,
    method::ColoringMethod,
    order::ColoringOrder;
    verbose::Bool=false    
)

ColPackColoring(
    M::SparseMatrixCSC,
    method::ColoringMethod,
    order::ColoringOrder;
    verbose::Bool=false    
)

Perform the coloring of a matrix that is either given directly or read from a file.

The users needs to specify a coloring method and an order on the vertices.

See also

ColoringMethod

ColoringOrder

ColPack.get_colors — Method

get_colors(coloring::ColPackColoring; verbose=false)

Retrieve the colors from a ColPackColoring as a vector of integers.

ColPack.matrix2adjmatrix — Method

matrix2adjmatrix(M::AbstractMatrix; partition_by_rows=true)

Create an adjacency matrix between the rows or between the columns of M, depending on whether partition_by_rows is true or false.

Coloring method

ColPack.ColoringMethod — Type

ColoringMethod

Represent a ColPack-compatible coloring method.

Fields

method::String

Constructors

d1_coloring()

d2_coloring()

acyclic_coloring()

star_coloring()

ColPack.acyclic_coloring — Method

acyclic_coloring()

Shortcut for ColoringMethod("ACYCLIC").

ColPack.d1_coloring — Method

d1_coloring()

Shortcut for ColoringMethod("DISTANCE_ONE").

ColPack.d2_coloring — Method

d2_coloring()

Shortcut for ColoringMethod("DISTANCE_TWO").

ColPack.star_coloring — Method

star_coloring()

Shortcut for ColoringMethod("STAR").

Coloring order

ColPack.ColoringOrder — Type

ColoringOrder

Represent a ColPack-compatible coloring order.

Fields

order::String

Constructors

natural_ordering()

largest_first_ordering()

dynamic_largest_first_ordering()

distance_two_largest_first_ordering()

smallest_last_ordering()

distance_two_smallest_last_ordering()

incidence_degree_ordering()

distance_two_incidence_degree_ordering()

random_ordering()

ColPack.distance_two_incidence_degree_ordering — Method

distance_two_incidence_degree_ordering()

Shortcut for ColoringOrder("DISTANCE_TWO_INCIDENCE_DEGREE").

ColPack.distance_two_largest_first_ordering — Method

distance_two_largest_first_ordering()

Shortcut for ColoringOrder("DISTANCE_TWO_LARGEST_FIRST").

ColPack.distance_two_smallest_last_ordering — Method

distance_two_smallest_last_ordering()

Shortcut for ColoringOrder("DISTANCE_TWO_SMALLEST_LAST").

ColPack.dynamic_largest_first_ordering — Method

dynamic_largest_first_ordering()

Shortcut for ColoringOrder("DYNAMIC_LARGEST_FIRST").

ColPack.incidence_degree_ordering — Method

incidence_degree_ordering()

Shortcut for ColoringOrder("INCIDENCE_DEGREE").

ColPack.largest_first_ordering — Method

largest_first_ordering()

Shortcut for ColoringOrder("LARGEST_FIRST").

ColPack.natural_ordering — Method

natural_ordering()

Shortcut for ColoringOrder("NATURAL").

ColPack.random_ordering — Method

random_ordering()

Shortcut for ColoringOrder("RANDOM").

ColPack.smallest_last_ordering — Method

smallest_last_ordering()

Shortcut for ColoringOrder("SMALLEST_LAST").

Internals

ColPack._cols_by_rows — Method

    _cols_by_rows(rows_index, cols_index)

Returns a vector of rows where each row contains a vector of its column indices.

ColPack._rows_by_cols — Method

    _rows_by_cols(rows_index, cols_index)

Returns a vector of columns where each column contains a vector of its row indices.